In search, bool keys are matched only for specific values.
Previously, if an ordinary term was given in a generic search, a zettel
with a boolean meta value of 'true' was matched most likely. Since
a boolean value is true, if it does not start with '0', 'f', 'F', 'N',
and 'n', many search terms that doe not start with theses characters
will produce a match for a true value.

This behaviour is still enabled for filtering, bit not for generic
search any more. In this case, it is tested, whether a search term
starts with one of the above characters or with onew of these: '1', 't',
'T', 'Y', or 'y'. Otherweise boolean keys are ignored.

FossilOrigin-Name: bb905997a4ce86508c103bf870b9a493383ebf969afc39b4f0fd87fa961a286b